refactor: move AgendaItemUi to separate file
